Fix for finding the default SDK in some scenarios
When trying to find the default SDK and there isn't a global.json
the default SDK finder sometimes throw a ArgumentOutOfRange exception
because the line breaks between sections in the output from --info
is missing.

This commit fixes one of those scenarios while keeping the original
way of finding the SDK intact.